# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AERANERD API",
        "description": "API para gerenciamento de áreas, matérias e escritores do sistema AERANERD",
        "contact": {
            "email": "contato@aeranerd.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/videos": {
            "get": {
                "tags": [
                    "Vídeos"
                ],
                "summary": "Listar todos os vídeos",
                "description": "Retorna uma lista paginada de todos os vídeos disponíveis",
                "operationId": "91a3a5e7add32d9ee4fb5ad638e10d66",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vídeos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Video"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vídeos"
                ],
                "summary": "Criar um novo vídeo",
                "description": "Cria um novo vídeo no sistema",
                "operationId": "de7b3eb3f2fce49c828dae36fff21116",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vchr_Titulo",
                                    "vchr_Description",
                                    "vchr_YTThumbDefault",
                                    "vchr_YTThumbMedium",
                                    "vchr_YTThumbHigh",
                                    "vchr_Embed",
                                    "vchr_ChannelId",
                                    "vchr_tags"
                                ],
                                "properties": {
                                    "int_IdMateria": {
                                        "description": "ID da matéria associada",
                                        "type": "integer"
                                    },
                                    "vchr_VideoId": {
                                        "description": "ID do vídeo no YouTube",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "vchr_LinkVideo": {
                                        "description": "Link do vídeo",
                                        "type": "string",
                                        "maxLength": 300
                                    },
                                    "int_IdArea": {
                                        "description": "ID da área",
                                        "type": "integer"
                                    },
                                    "vchr_Titulo": {
                                        "description": "Título do vídeo",
                                        "type": "string",
                                        "maxLength": 400
                                    },
                                    "vchr_Description": {
                                        "description": "Descrição do vídeo",
                                        "type": "string"
                                    },
                                    "vchr_YTThumbDefault": {
                                        "description": "Thumbnail padrão do YouTube",
                                        "type": "string",
                                        "maxLength": 300
                                    },
                                    "vchr_YTThumbMedium": {
                                        "description": "Thumbnail média do YouTube",
                                        "type": "string",
                                        "maxLength": 300
                                    },
                                    "vchr_YTThumbHigh": {
                                        "description": "Thumbnail alta do YouTube",
                                        "type": "string",
                                        "maxLength": 300
                                    },
                                    "vchr_Embed": {
                                        "description": "Código embed do vídeo",
                                        "type": "string"
                                    },
                                    "vchr_ChannelId": {
                                        "description": "ID do canal do YouTube",
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "vchr_tags": {
                                        "description": "Tags do vídeo",
                                        "type": "string",
                                        "maxLength": 400
                                    },
                                    "dt_Publicado": {
                                        "description": "Data de publicação no YouTube",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vídeo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/videos/youtube/{videoId}": {
            "get": {
                "tags": [
                    "Vídeos"
                ],
                "summary": "Buscar vídeo por ID do YouTube",
                "description": "Retorna um vídeo específico pelo seu ID do YouTube",
                "operationId": "f74436dcd829f942ac4dda281eed5292",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "description": "ID do vídeo no YouTube",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vídeo encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vídeo não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}